chat_module/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.0.3
openai>=1.3.5
python-multipart>=0.0.6
httpx>=0.25.2
playwright>=1.40.0
tavily-python>=0.2.2
Pillow>=10.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
google-generativeai>=0.8.0
pypdf>=4.0.0
